Rename InstallSection to GenericInstall and move it to common directory
