+++ widgets/common/src/GenericDeployModal.tsx
@@ -4,7 +4,7 @@ import MissingSecretsError from './MissingSecretsError';
 import AccordionSectionWithDivider from './AccordionSectionWithDivider';
 import DeplomentInputsSection from './deployModal/DeploymentInputsSection';
 import DeployModalActions from './deployModal/DeployModalActions';
-import InstallSection from './deployModal/GenericInstall';
+import InstallSection from './GenericInstall';
 import { createInstallFunction, isWorkflowName, getWorkflowName } from './deployModal/execution';
 
 const { i18n } = Stage;
